CLI: Windows Compatability and Enhancements
Known issues addressed:
#380 (if user was using the CLI)
#202
#180 (sort of)

Updates:
Added compatability for Windows to "get_terminal_size" as to prevent cli.py from crashing when downloading a stream.
For Linux and OSX users, the os.popen call for "stty size" has been replaced with a subprocess.check_output call. This tweak puts the "display_progress_bar" function into a different process, reducing download time.
If either attempt at getting the terminal geometry fails, the default values are returned as follows in an attempt to prevent an error: rows: 80, columns: 24.

Enhancements:
Added optional arguments for a target directory and a custom name for the downloaded stream.
If either argument isn't specified, the default action occurs for each (target is current working directory and name is the stream's original name, both of which are stream.download's defaults).