Separation from Master
Summary:
This branch solely contains components from the Master branch that allows users on the Windows operating system to use "cli.py" without it immediately crashing on runtime.

See the "Update" section in the "CLI: Windows Compatability and Enhancements" commit on my fork to view a more detailed description on what's changed.