Bypass YouTube whole download throttling
YouTube is throttling downloads without the Range header, which is used by browsers when streaming (for example with an implementation of the <video> HTML tag ) but not on whole downloads. We can easily bypass this, however, by just adding Range: bytes=0- to the request. The result is all the same bytes as before but in a much shorter time.