Fallback to index "signatureCipher"
It seems like YouTube recently changed the key name for "cipher" to
"signatureCipher" and PyTube has been failing because of this.

With this commit, PyTube will fallback to check for the value for
"signatureCipher" when the value for "cipher" is not available.
This is similar to how youtube-dl checks for this too
(in https://github.com/ytdl-org/youtube-dl/blob/228c1d685bb6d35b2c1a73e1adbc085c76da6b75/youtube_dl/extractor/youtube.py#L1977).

Fixes #641.