Fix for the new youtube playlist requests
Youtube changed their request and response format for playlists between 2021.03.02 and 2021.03.03.

This should also fix #949

Implements a new request function, because the new requests need to be post. This could technically be implemented into the get function, because "method" does not really influence the behaviour of urllib.request.Request, but the presence/absence of "data" does. To avoid confusion, its implemented in a new function.

Implements data and data encoding into _execute_request for the post requests.

Implemented content-type headers for the post requests, since the youtube servers are strict on content-type and if not added to the request headers this results in HTTPError [400]: Bad Request for every request.

Implemented data into the _execute_request and _build_continuation_url functions, which is required for the post requests.

Changed the format for the urls int the _build_continuation_url functions, which requires an api key. It was extracted from youtube playlist continuation requests on a browser, but it can also be found in a script on the initial playlist html, which contains "ytcfg.set". If necessary, extracting the api key from the initial html can be implemented.

Implemented a new format for the direct json responses from the youtube servers. The response is no longer a list and does not have the "response" key anymore.